+++ src/vmm/src/vmm_config/net.rs
@@ -10,7 +10,6 @@ use super::RateLimiterConfig;
 use crate::Error as VmmError;
 use devices::virtio::net::TapError;
 use devices::virtio::Net;
-use rate_limiter::{BucketUpdate, TokenBucket};
 use utils::net::mac::MacAddr;
 
 use serde::Deserialize;
@@ -59,50 +58,6 @@ pub struct NetworkInterfaceUpdateConfig {
     /// New TX rate limiter config. Only provided data will be updated. I.e. if any optional data
     /// is missing, it will not be nullified, but left unchanged.
     pub tx_rate_limiter: Option<RateLimiterConfig>,
-}
-
-macro_rules! get_bucket_update {
-    ($self:ident, $rate_limiter: ident, $metric: ident) => {{
-        match &$self.$rate_limiter {
-            Some(rl_cfg) => match rl_cfg.$metric {
-                // There is data to update.
-                Some(tb_cfg) => {
-                    TokenBucket::new(
-                        tb_cfg.size,
-                        tb_cfg.one_time_burst.unwrap_or(0),
-                        tb_cfg.refill_time,
-                    )
-                    // Updated active rate-limiter.
-                    .map(BucketUpdate::Update)
-                    // Updated/deactivated rate-limiter
-                    .unwrap_or(BucketUpdate::Disabled)
-                }
-                // No update to the rate-limiter.
-                None => BucketUpdate::None,
-            },
-            // No update to the rate-limiter.
-            None => BucketUpdate::None,
-        }
-    }};
-}
-
-impl NetworkInterfaceUpdateConfig {
-    /// Provides a `BucketUpdate` description for the RX bandwidth rate limiter.
-    pub fn rx_bytes(&self) -> BucketUpdate {
-        get_bucket_update!(self, rx_rate_limiter, bandwidth)
-    }
-    /// Provides a `BucketUpdate` description for the RX ops rate limiter.
-    pub fn rx_ops(&self) -> BucketUpdate {
-        get_bucket_update!(self, rx_rate_limiter, ops)
-    }
-    /// Provides a `BucketUpdate` description for the TX bandwidth rate limiter.
-    pub fn tx_bytes(&self) -> BucketUpdate {
-        get_bucket_update!(self, tx_rate_limiter, bandwidth)
-    }
-    /// Provides a `BucketUpdate` description for the TX ops rate limiter.
-    pub fn tx_ops(&self) -> BucketUpdate {
-        get_bucket_update!(self, tx_rate_limiter, ops)
-    }
 }
 
 /// Errors associated with `NetworkInterfaceConfig`.